requirements.txt
Flask==3.1.2
opencv-python-headless
deepface
numpy
Pillow
gunicorn
